Implement microphone_utils.py
